Rename reader module and file
