+++ src/lib.rs
@@ -1,8 +1,8 @@
 mod printer;
 pub use printer::pr_str;
 
-mod reader2;
-pub use reader2::{read_str, Error as ReaderError};
+mod reader;
+pub use reader::{read_str, Error as ReaderError};
 
 mod types;
 pub use types::{Atom, Bool, Comment, Complex, Float, Form, Integer, List, Map, Rational, Vector};
